[FEATURE] Introduce command to create local extension artefact
This commit adds a new `create-artefact` command. It can be used
to create a local artefact file (zip archive) of an extension. This
is basically the same which is already done in the `ter:publish`
command, except for the TER publish part.

The new command is especially useful to test custom configurations
of package exclude files. In addition, it's a useful helper in
cases where the TER REST API might not be accessible.