[FEATURE] Introduce command to create local extension artefact
* [FEATURE] Introduce command to create local extension artefact

This commit adds a new `create-artefact` command. It can be used
to create a local artefact file (zip archive) of an extension. This
is basically the same which is already done in the `ter:publish`
command, except for the TER publish part.

The new command is especially useful to test custom configurations
of package exclude files. In addition, it's a useful helper in
cases where the TER REST API might not be accessible.

* [TASK] Use different transaction path for `create-artefact` command

* [TASK] Include transaction path in exception message

* [TASK] Drop AbstractCommand in favor of CommandHelper utility class

* [TASK] Use admonitions and create custom section for packaging excludes