[DOCS] Simplify workflow examply in `README.md`
The `README.md` contains a GitHub workflow example
for publishing in on pushed tags and extracts from
a annotated tag content created with

```shell
git tag -a -m "some content"
```

filtering the git tag list for the extracted version
and telling to display max 10 lines for the content,
and extracting the tag prefix from the first line
using a `sed` regexp pattern.

`git` provides a simpler way using the `format` option,
which this change adopts in the aforementioned example.

To be precise, following:

```shell
git tag -n10 -l ${{ env.version }} | sed "s/^[0-9.]*[ ]*//g"
```

is replaced with

```shell
git tag -l ${{ env.version }} --format '%(contents)'
```

which displays only the tag content and removes the need
to extract the tag from the first line.